refactor: added joysticks event handlers for sdl2 and sdl1
