+++ include/core/apis/sdl1/event_handler.h
@@ -10,16 +10,17 @@
 #include "../../engine_properties/pause_gameloop_for_rendering.h"
 
 #include "../../../utils/keyboard.h" // For keyboard utils
-#include "../../../utils/joystick.h" // For joystick utils
 #include "../../../utils/mouse.h"    // For mouse utils
 
 #include "../../managers/event_processors.h" // For main event processors
 
+#include "joystick_event_handler.h"
+
 /**
  * @brief Sets the render delay value.
- * 
+ *
  * This macro utilizes SDL_Delay to pause execution for the given delay value.
- * 
+ *
  * @param delayValue The value for the render delay.
  */
 #define renderDelay(delayValue) SDL_Delay(delayValue)
@@ -53,12 +54,25 @@ bool initializeEventManagers()
         goto fail_initialization;
     }
 
+    if (!initializeEventManager(&gMouseEventManager, MOUSE_LAST))
+    {
+        fputs("ERROR: Cannot initialize mouse event manager", stderr);
+        goto fail_initialization;
+    }
+
+    if (!initializeJoystickEventHandler())
+    {
+        fputs("ERROR: Cannot initialize joystick events handler", stderr);
+        goto fail_initialization;
+    }
+
     goto success_initalization;
 
 fail_initialization:
     freeEventManager(&gKeyPressEventManager, true);
     freeEventManager(&gKeyReleaseEventManager, true);
     freeEventManager(&gMouseEventManager, true);
+    closeJoystickEventHandler();
     return false;
 
 success_initalization:
@@ -111,13 +125,13 @@ void processAllEvents()
             PROCESS_QUIT(event);
             break;
         }
-            // default:
-            //     handleJoystickEvents(&event);
-            //     break;
+        default:
+            handleJoystickEvents(&event);
+            break;
         }
     }
     // Trigger events for all joystick buttons still pressed
-    // TRIGGER_PRESSED_JOYSTICKS_BUTTONS();
+    TRIGGER_PRESSED_JOYSTICKS_BUTTONS();
 
     // Trigger events for all keys still pressed
     TRIGGER_PRESSED_KEYS();
@@ -151,10 +165,7 @@ void closeEventManagers()
     freeEventManager(&gMouseEventManager, true);
     freeEventManager(&gKeyPressEventManager, true);
     freeEventManager(&gKeyReleaseEventManager, true);
-
-    // closeJoystickManager();
-    // if (gJoyStickEventManager.listsCount > 0)
-    // freeEventManager(&gJoyStickEventManager, true);
+    closeJoystickEventHandler();
 }
 
 #endif
