+++ include/core/constants/joystick_max_axis_value.h
@@ -1,0 +1,7 @@
+#ifndef MAX_JOYSTICK_AXIS_VALUE_H
+#define MAX_JOYSTICK_AXIS_VALUE_H
+
+// Define maximum axis value for normalization. The range is adjusted by 0.5 to fix the range to [-32767.5, 32767.5]
+#define MAX_AXIS_VALUE 32767.5
+
+#endif
